requirements.txt
langchain
dotenv
requests
streamlit
langchain_openai
